add response class, need to integrate it to client class
