+++ nginx_docker/nginx.conf
@@ -30,7 +30,12 @@ http {
             set $post_data_handled 0;
             if ($request_method = POST) {
                 set $post_data_handled 1;
-            }
+            }       
+            add_header X-Debug-Webserv-Type "$http_webserv_type" always;
+            add_header X-Debug-Request-Method $request_method always;
+            if ($http_webserv_type = "bonjour"){
+                return 200 'POST request received on default server with webserv_type = bonjour';
+                }
 
             # Return response for POST requests
             if ($post_data_handled) {
@@ -65,6 +70,11 @@ http {
         error_log   /var/log/nginx/error.log debug;
 
         location /html2 {
+                   add_header X-Debug-Webserv-Type $http_webserv_type;
+        add_header X-Debug-Request-Method $request_method;
+            if ($http_webserv_type = "bonjour"){
+                return 200 'POST request received on POST server with webserv_type = bonjour';
+                }
             # Log all GET and POST requests to a file
             access_log /var/log/nginx/endpoint_access.log combined;
